fix: batch-strategy
