+++ packages/bridge-status-controller/src/utils/transaction.ts
@@ -381,7 +381,6 @@ export const getAddTransactionBatchParams = async ({
       'Failed to submit cross-chain swap batch transaction: unknown account in trade data',
     );
   }
-
   const hexChainId = formatChainIdToHex(trade.chainId);
   const networkClientId = getNetworkClientIdByChainId(messenger, hexChainId);
